convert a string to an integer

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -9,7 +9,7 @@
 int _atoi(char *s)
 {
 	int sign = 1;
-	unsigned long int result = 0;
+	unsigned int result = 0;
 	int i = 0;
 	int digit = 0;
